Add enhancements from local-development/src/kafka~

+++ coordinator-common/src/main/java/org/apache/kafka/coordinator/common/runtime/PartitionWriter.java
@@ -105,7 +105,7 @@ public interface PartitionWriter {
         String transactionalId,
         long producerId,
         short producerEpoch,
-        int apiVersion
+        short apiVersion
     ) throws KafkaException;
 
     /**
